// Richbourn-Dashboard/src/app/core/models/ledger.model.ts
export interface LedgerModel {
    GLCode: string;
    Location: string;
    Series: string;
    Name : String;
    GroupID : String;
    id: any;
    SubGrpID:number;
    SlNo: any;
    Alias: String;
    GlType:number;
    CityId:number;
    CityName: String;
    State: String;
    GroupDesc: String;
    GlBillingAdd: String;
    ZipCode: String;
    SubGroup: String;
    LedgerType: String;
    AcType: String;
    AcHolder: String;
    AccountNo: String;
    IfscCode: String;
    Country: String;
    TelePhone: String;
    Email: String;
    ContactPerson: String;

}